requirements.txt
asgiref==3.5.0
autopep8==1.6.0
bleach==4.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.3
django-axes==5.35.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-ipware==4.0.2
django-summernote==0.8.20.0
gunicorn==20.1.0
idna==3.3
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pyparsing==3.0.7
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
urllib3==1.26.10
webencodings==0.5.1
whitenoise==6.2.0
